Move geonodes_enum function to operators.py for easier refactoring later on
