+++ gn_items.py
@@ -210,15 +210,3 @@ def geonodes_node_items(context):
             )
             item_list.append(gn_nodeitem)
     return item_list
-
-def geonodes_enum(context):
-    items = [(
-        f'{node.label}',
-        f'{node.label} ({"".join(word[0] for word in node.label.split())})',
-        ""
-    ) for node in geonodes_node_items(context)]
-
-    items.append(("Simulation Zone", "Simulation Zone (SZ)", ""))
-    items.append(("Repeat Zone", "Repeat Zone (RZ)", ""))
-
-    return items
